requirements.txt
asgiref==3.3.1
autopep8==1.5.7
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-jazzmin==2.6.0
django-js-asset==1.2.2
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.6
pycodestyle==2.7.0
PyJWT==2.8.0
pytz==2021.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==5.2.0
